[refactor] biid 변수명 등 bid로 수정
