+++ client/src/Navigator/MainTab/designerTabStack.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 
-import BiidTabScreen from './User/biidTabScreen';
+import BiidTabScreen from './User/bidTabScreen';
 import SearchTabScreen from './User/searchTabScreen';
 import HistoryTabScreen from './User/historyTabScreen';
 import MyPageTabScreen from './User/myPageTabScreen';
